Add persisting user if logged in
